+++ src/main/java/com/jerry/mekextras/common/tier/TTier.java
@@ -6,19 +6,19 @@ import mekanism.common.tier.TubeTier;
 public class TTier {
     public static long getTubePullAmount(TubeTier tier) {
         return switch (tier) {
-            case BASIC -> ExtraConfig.extraTierConfig.absolutePressurizedTubePullAmount.get().longValue();
-            case ADVANCED -> ExtraConfig.extraTierConfig.supremePressurizedTubePullAmount.get().longValue();
-            case ELITE -> ExtraConfig.extraTierConfig.cosmicPressurizedTubePullAmount.get().longValue();
-            case ULTIMATE -> ExtraConfig.extraTierConfig.infinitePressurizedTubePullAmount.get().longValue();
+            case BASIC -> ExtraConfig.extraTierConfig.absolutePressurizedTubePullAmount.get();
+            case ADVANCED -> ExtraConfig.extraTierConfig.supremePressurizedTubePullAmount.get();
+            case ELITE -> ExtraConfig.extraTierConfig.cosmicPressurizedTubePullAmount.get();
+            case ULTIMATE -> ExtraConfig.extraTierConfig.infinitePressurizedTubePullAmount.get();
         };
     }
 
     public static long getTubeCapacity(TubeTier tier) {
         return switch (tier) {
-            case BASIC -> ExtraConfig.extraTierConfig.absolutePressurizedTubeCapacity.get().longValue();
-            case ADVANCED -> ExtraConfig.extraTierConfig.supremePressurizedTubeCapacity.get().longValue();
-            case ELITE -> ExtraConfig.extraTierConfig.cosmicPressurizedTubeCapacity.get().longValue();
-            case ULTIMATE -> ExtraConfig.extraTierConfig.infinitePressurizedTubeCapacity.get().longValue();
+            case BASIC -> ExtraConfig.extraTierConfig.absolutePressurizedTubeCapacity.get();
+            case ADVANCED -> ExtraConfig.extraTierConfig.supremePressurizedTubeCapacity.get();
+            case ELITE -> ExtraConfig.extraTierConfig.cosmicPressurizedTubeCapacity.get();
+            case ULTIMATE -> ExtraConfig.extraTierConfig.infinitePressurizedTubeCapacity.get();
         };
     }
 }
